+++ Changelog.md
@@ -10,7 +10,7 @@
 
 ### Enhancements
 
-* Added `maxRate` combinator to control the yield rate of a stream.
+* Added `yieldRate` combinator to control the yield rate of a stream.
 * Add `foldl1'`, `foldr1`, `intersperseM`, `find`, `lookup`, `and`, `or`,
   `findIndices`, `findIndex`, `elemIndices`, `elemIndex`, `init` to Prelude